Transformers.Basic

Basic functionality of Transformers.jl, provide the Transformer encoder/decoder implementation and other convenient function.

Transformer

The Transformer and TransformerDecoder is the encoder and decoder block of the origin paper, and they are all implement as the regular Flux Layer, so you can treat them just as the Dense layer. See the docstring for the argument. However, for the sequence data input, we usually have a 3 dimensional input of shape (hidden size, sequence length, batch size) instead of just (hidden size, batch size). Therefore, we implement both 2d & 3d operation according to the input type (The N of Array{T, N}). We are able to handle both input of shape (hidden size, sequence length, batch size) and (hidden size, sequence length) for the case with only 1 input.

using Transfomers

m = Transformer(512, 8, 64, 2048) #define a Transformer block with 8 head and 64 neuron for each head
x = randn(512, 30, 3) #fake data of length 30

y = m(x)

Positionwise

For the sequential task, we need to handle the 3 dimensional input. However, most of the layer in Flux only support input with shape (hidden size, batch size). In order to tackle this problem, we implement the Positionwise helper function that is almost the same as Flux.Chain but it will run the model position-wisely. (internally it just reshape the input to 2d and apply the model then reshape back).

using Transformers
using Flux

m = Positionwise(Dense(10, 5), Dense(5, 2), softmax)
x = randn(10, 30, 3)

y = m(x)

# which is equivalent to 
# 
# m = Chain(Dense(10, 5), Dense(5, 2), softmax)
# x1 = randn(10, 30)
# x2 = randn(10, 30)
# x3 = randn(10, 30)
# y = cat(m(x1), m(x2), m(x3); dims=3)

PositionEmbedding

We implement two kinds of position embedding, one is based on the sin/cos function (mentioned in the paper, attention is all you need). Another one is just like regular word embedding but with the position index. The first argument is the size. Since the position embedding is only related to the length of the input ( we use size(input, 2) as the length), the return value of the layer will be the embedding of the given length without duplicate to the batch size. you can/should use broadcast add to get the desired output.

# sin/cos based position embedding which is not trainable
pe = PositionEmbedding(10) # or PositionEmbedding(10; trainable = false)

# trainable position embedding need to specify the maximum length
pe = PositionEmbedding(10, 1024; trainable = true)

x = randn(Float32, 10, 6, 3) #fake data of shape (10, length = 6, batched_size = 3)

e = pe(x) #get the position embedding
y = x .+ e # add the position embedding to each sample

API Reference

Transformers.Basic.CompositeEmbedding — Method

CompositeEmbedding(;postprocessor=identity, es...)

composite several embedding into one embedding according the aggregate methods and apply postprocessor on it.

Transformers.Basic.Embed — Type

Embed(size::Int, vocab_size::Int)

The Embedding Layer, size is the hidden size. vocab_size is the number of the vocabulary. Just a wrapper for embedding matrix.

Transformers.Basic.PositionEmbedding — Type

PositionEmbedding(size::Int, max_len::Int = 1024; trainable::Bool = false)

The position embedding layer. size is the number of neuron. max_len is the maximum acceptable length of input. If is not trainable, max_len will dynamically adjust to the longest input length. If trainable, use a random init embedding value, otherwise use a sin/cos position encoding.

Transformers.Basic.Positionwise — Type

Positionwise(layers)

just like Flux.Chain, but reshape input to 2d and reshape back when output. Work exactly the same as Flux.Chain when input is 2d array.

Transformers.Basic.Transformer — Method

Transformer(size::Int, head::Int, ps::Int;
            future::Bool = true, act = relu, pdrop = 0.1)
Transformer(size::Int, head::Int, hs::Int, ps::Int;
            future::Bool = true, act = relu, pdrop = 0.1)

Transformer layer.

size is the input size. if hs is not specify, use div(size, head) as the hidden size of multi-head attention. ps is the hidden size & act is the activation function of the positionwise feedforward layer. When future is false, the k-th token can't see the j-th tokens where j > k. pdrop is the dropout rate.

Transformers.Basic.TransformerDecoder — Method

TransformerDecoder(size::Int, head::Int, ps::Int; act = relu, pdrop = 0.1)
TransformerDecoder(size::Int, head::Int, hs::Int, ps::Int; act = relu, pdrop = 0.1)

TransformerDecoder layer. Decode the value from a Encoder.

size is the input size. if hs is not specify, use div(size, head) as the hidden size of multi-head attention. ps is the hidden size & act is the activation function of the positionwise feedforward layer. pdrop is the dropout rate.

Transformers.Basic.TransformerModel — Type

TransformerModel(embed::AbstractEmbed, transformers::AbstractTransformer)
TransformerModel(
                  embed::AbstractEmbed,
                  transformers::AbstractTransformer,
                  classifier
                 )

a structure for putting everything together

Transformers.Basic.Vocabulary — Type

Vocabulary{T}(voc::Vector{T}, unk::T) where T

struct for holding the vocabulary list to encode/decode input tokens.

Transformers.Basic.Vocabulary — Method

(vocab::Vocabulary)(x)

encode the given data to the index encoding.

Transformers.Basic.clear_classifier — Method

clear_classifier(model::TransformerModel)

return a new TransformerModel without classifier.

Transformers.Basic.encode — Method

encode(vocab::Vocabulary, x)

encode the given data to the index encoding.

Transformers.Basic.gather — Method

gather(w::AbstractMatrix{T}, xs) where

getting vector at the given indices, xs is a array of indices(Int type).

Transformers.Basic.gather — Method

gather(w::AbstractMatrix{T}, xs::OneHotArray) where

getting vector at the given onehot encoding.

Transformers.Basic.gather — Method

gather(w::AbstractArray{T}, xs) where

getting vector at the given indices, xs is a array of cartesian indices(Tuple{Int} type).

Transformers.Basic.getmask — Method

getmask(ls::Container{<:Container})

get the mask for batched data.

Transformers.Basic.getmask — Method

getmask(m1::A, m2::A) where A <: Abstract3DTensor

get the mask for the covariance matrix.

Transformers.Basic.logcrossentropy — Method

compute the cross entropy with mask where p is already the log(p)

Transformers.Basic.logcrossentropy — Method

compute the cross entropy where p is already the log(p)

Transformers.Basic.logkldivergence — Method

compute the kl divergence with mask where p is already the log(p)

Transformers.Basic.logkldivergence — Method

compute the kl divergence where p is already the log(p)

Transformers.Basic.onehot2indices — Method

turn one hot encoding to indices

Transformers.Basic.set_classifier — Method

set_classifier(model::TransformerModel, classifier)

return a new TransformerModel whose classifier is set to classifier.

Transformers.Basic.@toNd — Macro

@toNd f(x, y, z...; a=a, b=b, c=c...) n

macro for calling 2-d array function on N-d array by reshape input with reshape(x, size(x, 1), :) and reshape back with reshape(out, :, input[n][2:end]...) where n is the n-th input(default=1).

Transformers.Basic.MultiheadAttention — Method

MultiheadAttention(head::Int, is::Int, hs::Int, os::Int;
                   future::Bool=true, pdrop = 0.1)

Multihead dot product Attention Layer, head is the number of head, is is the input size, hs is the hidden size of input projection layer of each head, os is the output size. When future is false, the k-th token can't see tokens at > k. pdrop is the dropout rate.

Transformers.Basic.PwFFN — Type

just a wrapper for two dense layer.