Transformers.BidirectionalEncoder

Implementation of BERT model

Get Pretrain

One of the pleasant feature BERT give us is that it make using a large pre-trained model on a language task possible. To use a pretrained model with Transformers.jl, you can either:

Use the public released weight with pretrain"" as show in the pretrain section

Convert the TensorFlow checkpoint file with BidirectionalEncoder.tfckpt2bson(need TensorFlow.jl installed). As long as the checkpoint is produced by the origin bert released by google, it should work without any issue.

Finetuning

After getting the pretrain weight in Julia. We are going to finetune the model on your dataset. The bert model is also a Flux layer, so training bert is just like training other Flux model (i.e. all the usage should be compatibled with Flux's API)

API reference

Transformers.BidirectionalEncoder.Bert — Type

Bert(size::Int, head::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)
Bert(size::Int, head::Int, hs::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)

the Bidirectional Encoder Representations from Transformer(BERT) model.

(bert::Bert)(x, mask=nothing; all::Bool=false)

eval the bert layer on input x. If length mask is given (in shape (1, seqlen, batchsize)), mask the attention with getmask(mask, mask). Moreover, set all to true to get all outputs of each transformer layer.

Transformers.BidirectionalEncoder.WordPiece — Type

WordPiece(vocab::Vector{String}, unk::String = "[UNK]"; max_char::Int=200)

WordPiece implementation.

(wp::WordPiece)(token)

split given token.

(wp::WordPiece)(tokens::Vector{String})

split given tokens

(wp::WordPiece)(type, tokens::Vector{String})

split given tokens, if type is Int, return pieces indices instead of strings pieces.

(wp::WordPiece)(tks::Vector{T}, token::String)

split given token and add result to tks. if T is Int, add indices instead of strings pieces.

Transformers.BidirectionalEncoder.BertTextEncoder — Type

struct BertTextEncoder{T<:AbstractTokenizer,
                       V<:AbstractVocabulary{String},
                       P} <: AbstractTextEncoder
  tokenizer::T
  vocab::V
  process::P
  startsym::String
  endsym::String
  trunc::Union{Nothing, Int}
end

The text encoder for Bert model. Taking a tokenizer, vocabulary, and a processing function, configured with a start symbol, an end symbol, and a maximum length.

BertTextEncoder(bert_tokenizer, wordpiece, process;
                startsym = "[CLS]", endsym = "[SEP]", trunc = nothing)

There are two tokenizer supported (bert_cased_tokenizer and bert_uncased_tokenizer). process can be omitted, then a predefined processing pipeline will be used.

BertTextEncoder(f, bertenc::BertTextEncoder)

Take a bert text encoder and create a new bert text encoder with same configuration except the processing function. f is a function that take the encoder and return a new process function. This is useful for changing part of the procssing function.

Example

julia> wordpiece = pretrain"bert-cased_L-12_H-768_A-12:wordpiece"
[ Info: loading pretrain bert model: cased_L-12_H-768_A-12.tfbson wordpiece
WordPiece(vocab_size=28996, unk=[UNK], max_char=200)

julia> bertenc = BertTextEncoder(bert_cased_tokenizer, wordpiece; trunc=5)
BertTextEncoder(
├─ TextTokenizer(WordPieceTokenization(bert_cased_tokenizer, WordPiece(vocab_size=28996, unk=[UNK], max_char=200))),
├─ vocab = Vocab{String, SizedArray}(size = 28996, unk = [UNK], unki = 101),
├─ startsym = [CLS],
├─ endsym = [SEP],
├─ trunc = 5,
└─ process = Pipelines:
  ╰─ target[tok] := nestedcall(string_getvalue, source)
  ╰─ target[tok] := with_firsthead_tail([CLS], [SEP])(target.tok)
  ╰─ target[(tok, segment)] := segment_and_concat(target.tok)
  ╰─ target[trunc_tok] := trunc_and_pad(5, [UNK])(target.tok)
  ╰─ target[trunc_len] := nestedmaxlength(target.trunc_tok)
  ╰─ target[mask] := getmask(target.tok, target.trunc_len)
  ╰─ target[tok] := nested2batch(target.trunc_tok)
  ╰─ target[segment] := (nested2batch ∘ trunc_and_pad(5, 1))(target.segment)
  ╰─ target[input] := (NamedTuple{(:tok, :segment)} ∘ tuple)(target.tok, target.segment)
  ╰─ target := (target.input, target.mask)
)

# take the first 3 pipeline and get the result
julia> BertTextEncoder(bertenc) do enc
           Pipelines(enc.process[1:3]) |> PipeGet{(:tok, :segment)}()
       end
BertTextEncoder(
├─ TextTokenizer(WordPieceTokenization(bert_cased_tokenizer, WordPiece(vocab_size=28996, unk=[UNK], max_char=200))),
├─ vocab = Vocab{String, SizedArray}(size = 28996, unk = [UNK], unki = 101),
├─ startsym = [CLS],
├─ endsym = [SEP],
├─ trunc = 5,
└─ process = Pipelines:
  ╰─ target[tok] := nestedcall(string_getvalue, source)
  ╰─ target[tok] := with_firsthead_tail([CLS], [SEP])(target.tok)
  ╰─ target[(tok, segment)] := segment_and_concat(target.tok)
  ╰─ target := (target.tok, target.segment)
)

Transformers.BidirectionalEncoder.bert_cased_tokenizer — Function

bert_cased_tokenizer(input)

google bert tokenizer which remain the case during tokenization. Recommended for multi-lingual data.

Transformers.BidirectionalEncoder.bert_uncased_tokenizer — Function

bert_uncased_tokenizer(input)

google bert tokenizer which do lower case on input before tokenization.

Transformers.BidirectionalEncoder.tfckpt2bson — Function

tfckpt2bson(path;
            raw=false,
            saveto="./",
            confname = "bert_config.json",
            ckptname = "bert_model.ckpt",
            vocabname = "vocab.txt")

turn google released bert format into BSON file. Set raw to true to remain the origin data format in bson.

Transformers.BidirectionalEncoder.load_bert_pretrain — Function

load_bert_pretrain(path, sym = :all)

Loading bert data/model from pretrain bson data. use sym to determine which data to load. possible value: :all(default), bert_model, wordpiece, tokenizer.

Transformers.BidirectionalEncoder.masklmloss — Function

masklmloss(embed::Embed{T}, transform,
           t::AbstractArray{T, N}, posis::AbstractArray{Tuple{Int,Int}}, labels) where {T,N}
masklmloss(embed::Embed{T}, transform, output_bias,
           t::AbstractArray{T, N}, posis::AbstractArray{Tuple{Int,Int}}, labels) where {T,N}

helper function for computing the maks language modeling loss. Performance transform(x) .+ output_bias where x is the mask specified by posis, then compute the similarity with embed.embedding and crossentropy between true labels.

Transformers.BidirectionalEncoder.bert_pretrain_task — Function

bert_pretrain_task(datachn::Channel, wordpiece::WordPiece;
                   buffer_size = 100,
                   channel_size = 100
                   wordpiece::WordPiece,
                   sentences_pool = sentences;
                   start_token = "[CLS]",
                   sep_token = "[SEP]",
                   mask_token = "[MASK]",
                   mask_ratio = 0.15,
                   real_token_ratio = 0.1,
                   random_token_ratio = 0.1,
                   whole_word_mask = false,
                   next_sentence_ratio = 0.5,
                   next_sentence = true,
                   tokenizer = tokenize,
                   istokenized = false,
                   return_real_sentence = false)

helper function to generate bert mask language modeling and next sentence prediction data. datachn is a Channel with input documents line by line.

Transformers.BidirectionalEncoder.recursive_readdir — Function

recursive_readdir(path::AbstractString="./")

recursive read all file from a dir. return a list of filenames.