Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4423680 bytes (1%) 8830976 bytes (2%) 13237841 bytes (3%) 17645568 bytes (4%) 22052231 bytes (5%) 26476544 bytes (6%) 30883376 bytes (7%) 35290672 bytes (8%) 39697356 bytes (9%) 44104132 bytes (10%) 48513024 bytes (11%) 52934171 bytes (12%) 57340481 bytes (13%) 61747197 bytes (14%) 66154386 bytes (15%) 70577232 bytes (16%) 74983497 bytes (17%) 79396864 bytes (18%) 83803531 bytes (19%) 88213173 bytes (20%) 92619776 bytes (21%) 97026171 bytes (22%) 101449220 bytes (23%) 105856516 bytes (24%) 110262939 bytes (25%) 114674321 bytes (26%) 119081617 bytes (27%) 123488172 bytes (28%) 127907369 bytes (29%) 132321424 bytes (30%) 136722602 bytes (31%) 141131147 bytes (32%) 145542862 bytes (33%) 149950158 bytes (34%) 154367325 bytes (35%) 158774621 bytes (36%) 163180915 bytes (37%) 167604595 bytes (38%) 172016547 bytes (39%) 176422850 bytes (40%) 180825025 bytes (41%) 185231360 bytes (42%) 189655040 bytes (43%) 194060760 bytes (44%) 198467165 bytes (45%) 202879900 bytes (46%) 207286312 bytes (47%) 211692879 bytes (48%) 216116559 bytes (49%) 220523118 bytes (50%) 224932425 bytes (51%) 229338771 bytes (52%) 233746804 bytes (53%) 238162507 bytes (54%) 242569803 bytes (55%) 246981923 bytes (56%) 251388633 bytes (57%) 255795929 bytes (58%) 260218581 bytes (59%) 264624975 bytes (60%) 269037568 bytes (61%) 273443879 bytes (62%) 277850410 bytes (63%) 282257706 bytes (64%) 286680699 bytes (65%) 291090864 bytes (66%) 295498160 bytes (67%) 299902281 bytes (68%) 304308941 bytes (69%) 308726210 bytes (70%) 313132576 bytes (71%) 317555487 bytes (72%) 321961764 bytes (73%) 326361855 bytes (74%) 330777030 bytes (75%) 335184326 bytes (76%) 339590621 bytes (77%) 344006976 bytes (78%) 348414272 bytes (79%) 352829305 bytes (80%) 357235598 bytes (81%) 361642954 bytes (82%) 366062257 bytes (83%) 370469553 bytes (84%) 374873137 bytes (85%) 379295979 bytes (86%) 383699573 bytes (87%) 388105847 bytes (88%) 392529483 bytes (89%) 396936672 bytes (90%) 401343321 bytes (91%) 405750617 bytes (92%) 410154925 bytes (93%) 414578605 bytes (94%) 418984885 bytes (95%) 423391355 bytes (96%) 427798651 bytes (97%) 432215085 bytes (98%) 436622381 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14105919 bytes (3%) 18808832 bytes (4%) 23494656 bytes (5%) 28196864 bytes (6%) 32898358 bytes (7%) 37601280 bytes (8%) 42303488 bytes (9%) 46988363 bytes (10%) 51691520 bytes (11%) 56392955 bytes (12%) 61095936 bytes (13%) 65781340 bytes (14%) 70483968 bytes (15%) 75186176 bytes (16%) 79888384 bytes (17%) 84589861 bytes (18%) 89275995 bytes (19%) 93978624 bytes (20%) 98680832 bytes (21%) 103382092 bytes (22%) 108084300 bytes (23%) 112771072 bytes (24%) 117473280 bytes (25%) 122175488 bytes (26%) 126876856 bytes (27%) 131578041 bytes (28%) 136265728 bytes (29%) 140967936 bytes (30%) 145670144 bytes (31%) 150371331 bytes (32%) 155057048 bytes (33%) 159760384 bytes (34%) 164462592 bytes (35%) 169164800 bytes (36%) 173850624 bytes (37%) 178552832 bytes (38%) 183255040 bytes (39%) 187948410 bytes (40%) 192650618 bytes (41%) 197351863 bytes (42%) 202047488 bytes (43%) 206748915 bytes (44%) 211448933 bytes (45%) 216144486 bytes (46%) 220839936 bytes (47%) 225541395 bytes (48%) 230242624 bytes (49%) 234946560 bytes (50%) 239632384 bytes (51%) 244334592 bytes (52%) 249035865 bytes (53%) 253739008 bytes (54%) 258440283 bytes (55%) 263125311 bytes (56%) 267826509 bytes (57%) 272527694 bytes (58%) 277233664 bytes (59%) 281918498 bytes (60%) 286621696 bytes (61%) 291322884 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314817872 bytes (67%) 319520768 bytes (68%) 324222492 bytes (69%) 328907409 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357101240 bytes (76%) 361807872 bytes (77%) 366509056 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380603138 bytes (81%) 385304585 bytes (82%) 389990228 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413482901 bytes (88%) 418185109 bytes (89%) 422892723 bytes (90%) 427578547 bytes (91%) 432276764 bytes (92%) 436980469 bytes (93%) 441681882 bytes (94%) 446375522 bytes (95%) 451077730 bytes (96%) 455779831 bytes (97%) 460469173 bytes (98%) 465171381 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("((x, m) => x':(x, m)) => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), segment = Embed(768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm((768,)), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768, 768, tanh)
      masklm => (
        transform => Chain(Dense(768, 768, gelu), LayerNorm((768,)))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768, 2), logsoftmax)
    )
)TransformerModel{Gpt{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("x':x => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().