BERT

Transformers.BidirectionalEncoder

Implementation of BERT model

API reference

Transformers.BidirectionalEncoder.Bert — Type.

Bert(size::Int, head::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)
Bert(size::Int, head::Int, hs::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)

the Bidirectional Encoder Representations from Transformer(BERT) model.

(bert::Bert)(x, mask=nothing; all::Bool=false)

eval the bert layer on input x. If length mask is given (in shape (1, seqlen, batchsize)), mask the attention with getmask(mask, mask). Moreover, set all to true to get all outputs of each transformer layer.

source

Transformers.BidirectionalEncoder.WordPiece — Type.

WordPiece(vocab::Vector{String}, unk::String = "[UNK]"; max_char::Int=200)

WordPiece implementation.

(wp::WordPiece)(token)

split given token.

(wp::WordPiece)(tokens::Vector{String})

split given tokens

(wp::WordPiece)(type, tokens::Vector{String})

split given tokens, if type is Int, return pieces indices instead of strings pieces.

(wp::WordPiece)(tks::Vector{T}, token::String)

split given token and add result to tks. if T is Int, add indices instead of strings pieces.

source

Transformers.BidirectionalEncoder.bert_cased_tokenizer — Function.

bert_cased_tokenizer(input)

google bert tokenizer which remain the case during tokenization. Recommended for multi-lingual data.

source

Transformers.BidirectionalEncoder.bert_uncased_tokenizer — Function.

bert_uncased_tokenizer(input)

google bert tokenizer which do lower case on input before tokenization.

source

Transformers.BidirectionalEncoder.tfckpt2bson — Function.

tfckpt2bson(path; raw=false, saveto="./", confname = "bert_config.json", ckptname = "bert_model.ckpt", vocabname = "vocab.txt")

turn google released bert format into BSON file. Set raw to true to remain the origin data format in bson.

source

Transformers.BidirectionalEncoder.masklmloss — Function.

masklmloss(embed::Embed{T}, transform, t::AbstractArray{T, N}, posis::AbstractArray{Tuple{Int,Int}}, labels) where {T,N}
masklmloss(embed::Embed{T}, transform, output_bias, t::AbstractArray{T, N}, posis::AbstractArray{Tuple{Int,Int}}, labels) where {T,N}

helper function for computing the maks language modeling loss. Performance transform(x) .+ output_bias where x is specified by posis, then compute the similarity with embed.embedding and crossentropy between true labels.

source

Transformers.BidirectionalEncoder.bert_pretrain_task — Function.

bert_pretrain_task(datachn::Channel, wordpiece::WordPiece;
                   buffer_size = 100,
                   channel_size = 100
                   wordpiece::WordPiece,
                   sentences_pool = sentences;
                   start_token = "[CLS]",
                   sep_token = "[SEP]",
                   mask_token = "[MASK]",
                   mask_ratio = 0.15,
                   real_token_ratio = 0.1,
                   random_token_ratio = 0.1,
                   whole_word_mask = false,
                   next_sentence_ratio = 0.5,
                   next_sentence = true,
                   tokenizer = tokenize,
                   istokenized = false,
                   return_real_sentence = false)

helper function to generate bert mask language modeling and next sentence prediction data. datachn is a Channel with input documents line by line.

source

Transformers.BidirectionalEncoder.recursive_readdir — Function.

recursive_readdir(path::AbstractString="./")

recursive read all file from a dir. return a list of filenames.

source