GPT

Transformers.GenerativePreTrain

Implementation of gpt-1 model

API reference

Transformers.GenerativePreTrain.Gpt — Method.

Gpt(size::Int, head::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)
Gpt(size::Int, head::Int, hs::Int, ps::Int, layer::Int;
    act = gelu, pdrop = 0.1, attn_pdrop = 0.1)

the Generative Pretrained Transformer(GPT) model.

(gpt::Gpt)(x::T, mask=nothing; all::Bool=false)

eval the gpt layer on input x. If mask is given, mask the attention with mask. Moreover, set all to true to get all outputs of each transformer layer.

source

Transformers.GenerativePreTrain.lmloss — Method.

lmloss(embed, onehot, encoding, mask)

compute the language modeling loss for Gpt, onehot is the onehot array of the origin input sentence. encoding the output of Gpt, mask is the mask between input sentences.

source

Transformers.GenerativePreTrain.load_gpt_pretrain — Function.

load_gpt_pretrain(path::AbstractString, sym = :all; startsym="_start_", delisym="_delimiter_", clfsym="_classify_", unksym="<unk>")

load gpt data/model from pretrain bson data. use sym to determine which data to load. set <xxx>syms for setting special symbols in vocabulary. possible value: :all(default), gpt_model, bpe, vocab, tokenizer.

source

Transformers.GenerativePreTrain.gpt_tokenizer — Method.

gpt_tokenizer(x)

An alternative for origin tokenizer (spacy tokenizer) used in gpt model.

source

Transformers.GenerativePreTrain.openAInpy2bson — Method.

openAInpy2bson(path; raw=false, saveto="./", startsym="_start_", delisym="_delimiter_", clfsym="_classify_", unksym="<unk>")

turn openai released gpt format(npy) into BSON file. Set raw to true to remain the origin data format in bson.

source

Transformers.GenerativePreTrain.text_standardize — Method.

The function in the origin gpt code

fixes some issues the spacy tokenizer had on books corpus also does some whitespace standardization

source