Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4423680 bytes (1%) 8830976 bytes (2%) 13238272 bytes (3%) 17645568 bytes (4%) 22052864 bytes (5%) 26476544 bytes (6%) 30883840 bytes (7%) 35291136 bytes (8%) 39698432 bytes (9%) 44105728 bytes (10%) 48513024 bytes (11%) 52936704 bytes (12%) 57344000 bytes (13%) 61751296 bytes (14%) 66158592 bytes (15%) 70565888 bytes (16%) 74989568 bytes (17%) 79396864 bytes (18%) 83804160 bytes (19%) 88211456 bytes (20%) 92618752 bytes (21%) 97029437 bytes (22%) 101436733 bytes (23%) 105860413 bytes (24%) 110267709 bytes (25%) 114675005 bytes (26%) 119082301 bytes (27%) 123489597 bytes (28%) 127913277 bytes (29%) 132320573 bytes (30%) 136727869 bytes (31%) 141135165 bytes (32%) 145542461 bytes (33%) 149949757 bytes (34%) 154373437 bytes (35%) 158780733 bytes (36%) 163188029 bytes (37%) 167595325 bytes (38%) 172002621 bytes (39%) 176426301 bytes (40%) 180825946 bytes (41%) 185233242 bytes (42%) 189656922 bytes (43%) 194064218 bytes (44%) 198471514 bytes (45%) 202878810 bytes (46%) 207288143 bytes (47%) 211695439 bytes (48%) 216115262 bytes (49%) 220522558 bytes (50%) 224929854 bytes (51%) 229337150 bytes (52%) 233744446 bytes (53%) 238168126 bytes (54%) 242575422 bytes (55%) 246982718 bytes (56%) 251390014 bytes (57%) 255807371 bytes (58%) 260214667 bytes (59%) 264621963 bytes (60%) 269029259 bytes (61%) 273436555 bytes (62%) 277860235 bytes (63%) 282267531 bytes (64%) 286674827 bytes (65%) 291082123 bytes (66%) 295489419 bytes (67%) 299913099 bytes (68%) 304323325 bytes (69%) 308730621 bytes (70%) 313137917 bytes (71%) 317545213 bytes (72%) 321955457 bytes (73%) 326362753 bytes (74%) 330770049 bytes (75%) 335193729 bytes (76%) 339601025 bytes (77%) 344008321 bytes (78%) 348415617 bytes (79%) 352822913 bytes (80%) 357246593 bytes (81%) 361653889 bytes (82%) 366056837 bytes (83%) 370464133 bytes (84%) 374887813 bytes (85%) 379295109 bytes (86%) 383702405 bytes (87%) 388109701 bytes (88%) 392516997 bytes (89%) 396924293 bytes (90%) 401347973 bytes (91%) 405751908 bytes (92%) 410159204 bytes (93%) 414566415 bytes (94%) 418990095 bytes (95%) 423395820 bytes (96%) 427803116 bytes (97%) 432210412 bytes (98%) 436617708 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14106624 bytes (3%) 18808832 bytes (4%) 23494656 bytes (5%) 28196864 bytes (6%) 32899072 bytes (7%) 37601280 bytes (8%) 42303488 bytes (9%) 46989312 bytes (10%) 51691520 bytes (11%) 56393728 bytes (12%) 61095936 bytes (13%) 65781760 bytes (14%) 70483968 bytes (15%) 75186176 bytes (16%) 79888384 bytes (17%) 84590592 bytes (18%) 89276416 bytes (19%) 93978624 bytes (20%) 98680832 bytes (21%) 103383040 bytes (22%) 108068864 bytes (23%) 112771072 bytes (24%) 117473280 bytes (25%) 122175488 bytes (26%) 126877696 bytes (27%) 131563520 bytes (28%) 136265728 bytes (29%) 140967936 bytes (30%) 145670144 bytes (31%) 150372352 bytes (32%) 155058176 bytes (33%) 159760384 bytes (34%) 164462592 bytes (35%) 169164800 bytes (36%) 173850624 bytes (37%) 178552832 bytes (38%) 183255040 bytes (39%) 187957248 bytes (40%) 192659456 bytes (41%) 197345280 bytes (42%) 202047488 bytes (43%) 206749696 bytes (44%) 211451904 bytes (45%) 216137728 bytes (46%) 220839936 bytes (47%) 225542144 bytes (48%) 230244352 bytes (49%) 234946560 bytes (50%) 239632384 bytes (51%) 244334592 bytes (52%) 249036800 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206592 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380600320 bytes (81%) 385302528 bytes (82%) 389988352 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887424 bytes (90%) 427589632 bytes (91%) 432275456 bytes (92%) 436977664 bytes (93%) 441679872 bytes (94%) 446382080 bytes (95%) 451084288 bytes (96%) 455770112 bytes (97%) 460472320 bytes (98%) 465174528 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().