Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419717 bytes (1%) 8827013 bytes (2%) 13234309 bytes (3%) 17641605 bytes (4%) 22065285 bytes (5%) 26472581 bytes (6%) 30879877 bytes (7%) 35287173 bytes (8%) 39694469 bytes (9%) 44118149 bytes (10%) 48525445 bytes (11%) 52932741 bytes (12%) 57340037 bytes (13%) 61747333 bytes (14%) 66154629 bytes (15%) 70578309 bytes (16%) 74985605 bytes (17%) 79392901 bytes (18%) 83800197 bytes (19%) 88207493 bytes (20%) 92631173 bytes (21%) 97038469 bytes (22%) 101445765 bytes (23%) 105853061 bytes (24%) 110260357 bytes (25%) 114667653 bytes (26%) 119091333 bytes (27%) 123498629 bytes (28%) 127905925 bytes (29%) 132313221 bytes (30%) 136720517 bytes (31%) 141144197 bytes (32%) 145539072 bytes (33%) 149962752 bytes (34%) 154370048 bytes (35%) 158777344 bytes (36%) 163184640 bytes (37%) 167591936 bytes (38%) 172015616 bytes (39%) 176422912 bytes (40%) 180830208 bytes (41%) 185237504 bytes (42%) 189644800 bytes (43%) 194052096 bytes (44%) 198467899 bytes (45%) 202875195 bytes (46%) 207298875 bytes (47%) 211706171 bytes (48%) 216113467 bytes (49%) 220520763 bytes (50%) 224928059 bytes (51%) 229335355 bytes (52%) 233759035 bytes (53%) 238166331 bytes (54%) 242573627 bytes (55%) 246980923 bytes (56%) 251388219 bytes (57%) 255795515 bytes (58%) 260219195 bytes (59%) 264626491 bytes (60%) 269033787 bytes (61%) 273441083 bytes (62%) 277848379 bytes (63%) 282272059 bytes (64%) 286669864 bytes (65%) 291084082 bytes (66%) 295499116 bytes (67%) 299906412 bytes (68%) 304313708 bytes (69%) 308721004 bytes (70%) 313144684 bytes (71%) 317551980 bytes (72%) 321959276 bytes (73%) 326366572 bytes (74%) 330780456 bytes (75%) 335187752 bytes (76%) 339595048 bytes (77%) 344002344 bytes (78%) 348426024 bytes (79%) 352833320 bytes (80%) 357240616 bytes (81%) 361647912 bytes (82%) 366055208 bytes (83%) 370462504 bytes (84%) 374888754 bytes (85%) 379296050 bytes (86%) 383703346 bytes (87%) 388110642 bytes (88%) 392517938 bytes (89%) 396925234 bytes (90%) 401348914 bytes (91%) 405756210 bytes (92%) 410163506 bytes (93%) 414575279 bytes (94%) 418982575 bytes (95%) 423389871 bytes (96%) 427797167 bytes (97%) 432220847 bytes (98%) 436628143 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714646 bytes (1%) 9400470 bytes (2%) 14102678 bytes (3%) 18804886 bytes (4%) 23507094 bytes (5%) 28192918 bytes (6%) 32895126 bytes (7%) 37597334 bytes (8%) 42299542 bytes (9%) 47001750 bytes (10%) 51687574 bytes (11%) 56389782 bytes (12%) 61091990 bytes (13%) 65794198 bytes (14%) 70480022 bytes (15%) 75182230 bytes (16%) 79884438 bytes (17%) 84586646 bytes (18%) 89288854 bytes (19%) 93978624 bytes (20%) 98680832 bytes (21%) 103383040 bytes (22%) 108068864 bytes (23%) 112771072 bytes (24%) 117473280 bytes (25%) 122175488 bytes (26%) 126877696 bytes (27%) 131563520 bytes (28%) 136265728 bytes (29%) 140967936 bytes (30%) 145670144 bytes (31%) 150372352 bytes (32%) 155058176 bytes (33%) 159760384 bytes (34%) 164462592 bytes (35%) 169164800 bytes (36%) 173850624 bytes (37%) 178552832 bytes (38%) 183255040 bytes (39%) 187957248 bytes (40%) 192659456 bytes (41%) 197345280 bytes (42%) 202046491 bytes (43%) 206749569 bytes (44%) 211449189 bytes (45%) 216151397 bytes (46%) 220836809 bytes (47%) 225539017 bytes (48%) 230244352 bytes (49%) 234946560 bytes (50%) 239632384 bytes (51%) 244334592 bytes (52%) 249036800 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206592 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380600320 bytes (81%) 385302528 bytes (82%) 389988352 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887424 bytes (90%) 427589632 bytes (91%) 432275456 bytes (92%) 436977664 bytes (93%) 441679872 bytes (94%) 446382080 bytes (95%) 451084288 bytes (96%) 455770112 bytes (97%) 460472320 bytes (98%) 465174528 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().