Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4418916 bytes (1%) 8827007 bytes (2%) 13244808 bytes (3%) 17651303 bytes (4%) 22053459 bytes (5%) 26476736 bytes (6%) 30883228 bytes (7%) 35289055 bytes (8%) 39701769 bytes (9%) 44109020 bytes (10%) 48521063 bytes (11%) 52923537 bytes (12%) 57338777 bytes (13%) 61745312 bytes (14%) 66168226 bytes (15%) 70574781 bytes (16%) 74983424 bytes (17%) 79386829 bytes (18%) 83801778 bytes (19%) 88208272 bytes (20%) 92624717 bytes (21%) 97036821 bytes (22%) 101436323 bytes (23%) 105853932 bytes (24%) 110269440 bytes (25%) 114676583 bytes (26%) 119092411 bytes (27%) 123491239 bytes (28%) 127910799 bytes (29%) 132308824 bytes (30%) 136729427 bytes (31%) 141130485 bytes (32%) 145552709 bytes (33%) 149959206 bytes (34%) 154367015 bytes (35%) 158781396 bytes (36%) 163187623 bytes (37%) 167601862 bytes (38%) 172015608 bytes (39%) 176415562 bytes (40%) 180834840 bytes (41%) 185236054 bytes (42%) 189646412 bytes (43%) 194057216 bytes (44%) 198475776 bytes (45%) 202878084 bytes (46%) 207283050 bytes (47%) 211695328 bytes (48%) 216110970 bytes (49%) 220523338 bytes (50%) 224930383 bytes (51%) 229349760 bytes (52%) 233756236 bytes (53%) 238156800 bytes (54%) 242577408 bytes (55%) 246983651 bytes (56%) 251387659 bytes (57%) 255806120 bytes (58%) 260211712 bytes (59%) 264625738 bytes (60%) 269033035 bytes (61%) 273447082 bytes (62%) 277853956 bytes (63%) 282261596 bytes (64%) 286668095 bytes (65%) 291090962 bytes (66%) 295494954 bytes (67%) 299899148 bytes (68%) 304322065 bytes (69%) 308730313 bytes (70%) 313136427 bytes (71%) 317541243 bytes (72%) 321961992 bytes (73%) 326372339 bytes (74%) 330778164 bytes (75%) 335180542 bytes (76%) 339599569 bytes (77%) 344013690 bytes (78%) 348414086 bytes (79%) 352825622 bytes (80%) 357241411 bytes (81%) 361649933 bytes (82%) 366062222 bytes (83%) 370463247 bytes (84%) 374883612 bytes (85%) 379284179 bytes (86%) 383698349 bytes (87%) 388104852 bytes (88%) 392527887 bytes (89%) 396934542 bytes (90%) 401343756 bytes (91%) 405751079 bytes (92%) 410157632 bytes (93%) 414580562 bytes (94%) 418987209 bytes (95%) 423393864 bytes (96%) 427796917 bytes (97%) 432207738 bytes (98%) 436629366 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714490 bytes (1%) 9400314 bytes (2%) 14102522 bytes (3%) 18804730 bytes (4%) 23506938 bytes (5%) 28192762 bytes (6%) 32894970 bytes (7%) 37597178 bytes (8%) 42299386 bytes (9%) 47001594 bytes (10%) 51687418 bytes (11%) 56389626 bytes (12%) 61091834 bytes (13%) 65794042 bytes (14%) 70479866 bytes (15%) 75182074 bytes (16%) 79884282 bytes (17%) 84586490 bytes (18%) 89288698 bytes (19%) 93974522 bytes (20%) 98676730 bytes (21%) 103378938 bytes (22%) 108081146 bytes (23%) 112783354 bytes (24%) 117469178 bytes (25%) 122171386 bytes (26%) 126873594 bytes (27%) 131575802 bytes (28%) 136261626 bytes (29%) 140963834 bytes (30%) 145666042 bytes (31%) 150368250 bytes (32%) 155070458 bytes (33%) 159756282 bytes (34%) 164458490 bytes (35%) 169160698 bytes (36%) 173862906 bytes (37%) 178548730 bytes (38%) 183250938 bytes (39%) 187953146 bytes (40%) 192655354 bytes (41%) 197357562 bytes (42%) 202043386 bytes (43%) 206744238 bytes (44%) 211449131 bytes (45%) 216140135 bytes (46%) 220837634 bytes (47%) 225540349 bytes (48%) 230245564 bytes (49%) 234947772 bytes (50%) 239640648 bytes (51%) 244334592 bytes (52%) 249036800 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206592 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380600320 bytes (81%) 385302528 bytes (82%) 389988352 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887424 bytes (90%) 427589632 bytes (91%) 432275456 bytes (92%) 436977664 bytes (93%) 441679872 bytes (94%) 446382080 bytes (95%) 451084288 bytes (96%) 455770112 bytes (97%) 460472320 bytes (98%) 465174528 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().