Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419142 bytes (1%) 8826438 bytes (2%) 13233734 bytes (3%) 17657414 bytes (4%) 22064710 bytes (5%) 26472006 bytes (6%) 30879302 bytes (7%) 35286598 bytes (8%) 39693894 bytes (9%) 44117574 bytes (10%) 48524870 bytes (11%) 52932166 bytes (12%) 57339462 bytes (13%) 61746758 bytes (14%) 66154054 bytes (15%) 70577734 bytes (16%) 74985030 bytes (17%) 79392326 bytes (18%) 83799622 bytes (19%) 88206918 bytes (20%) 92630598 bytes (21%) 97037894 bytes (22%) 101445190 bytes (23%) 105852486 bytes (24%) 110259782 bytes (25%) 114667078 bytes (26%) 119090758 bytes (27%) 123498054 bytes (28%) 127905350 bytes (29%) 132312646 bytes (30%) 136719942 bytes (31%) 141143622 bytes (32%) 145550918 bytes (33%) 149958214 bytes (34%) 154365510 bytes (35%) 158772806 bytes (36%) 163180102 bytes (37%) 167603782 bytes (38%) 172011078 bytes (39%) 176418374 bytes (40%) 180825670 bytes (41%) 185232966 bytes (42%) 189656646 bytes (43%) 194063942 bytes (44%) 198471238 bytes (45%) 202878534 bytes (46%) 207285830 bytes (47%) 211693126 bytes (48%) 216112493 bytes (49%) 220519789 bytes (50%) 224927085 bytes (51%) 229334381 bytes (52%) 233758061 bytes (53%) 238165357 bytes (54%) 242572653 bytes (55%) 246979949 bytes (56%) 251387245 bytes (57%) 255810925 bytes (58%) 260218221 bytes (59%) 264625517 bytes (60%) 269032813 bytes (61%) 273440109 bytes (62%) 277847405 bytes (63%) 282271085 bytes (64%) 286678381 bytes (65%) 291085677 bytes (66%) 295492973 bytes (67%) 299900269 bytes (68%) 304323949 bytes (69%) 308731245 bytes (70%) 313138541 bytes (71%) 317545837 bytes (72%) 321953133 bytes (73%) 326360429 bytes (74%) 330784109 bytes (75%) 335191405 bytes (76%) 339598701 bytes (77%) 344005997 bytes (78%) 348413293 bytes (79%) 352836973 bytes (80%) 357244269 bytes (81%) 361651565 bytes (82%) 366058861 bytes (83%) 370466157 bytes (84%) 374873453 bytes (85%) 379297133 bytes (86%) 383704429 bytes (87%) 388111725 bytes (88%) 392519021 bytes (89%) 396926317 bytes (90%) 401349997 bytes (91%) 405757293 bytes (92%) 410164589 bytes (93%) 414571885 bytes (94%) 418979181 bytes (95%) 423386477 bytes (96%) 427810157 bytes (97%) 432217453 bytes (98%) 436624749 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14106624 bytes (3%) 18808832 bytes (4%) 23494656 bytes (5%) 28196864 bytes (6%) 32899072 bytes (7%) 37601280 bytes (8%) 42303488 bytes (9%) 46989312 bytes (10%) 51691149 bytes (11%) 56393357 bytes (12%) 61095565 bytes (13%) 65781389 bytes (14%) 70483597 bytes (15%) 75185805 bytes (16%) 79887381 bytes (17%) 84589589 bytes (18%) 89275413 bytes (19%) 93977621 bytes (20%) 98679829 bytes (21%) 103382037 bytes (22%) 108084245 bytes (23%) 112770069 bytes (24%) 117472277 bytes (25%) 122174485 bytes (26%) 126876693 bytes (27%) 131562517 bytes (28%) 136264725 bytes (29%) 140966933 bytes (30%) 145669141 bytes (31%) 150371349 bytes (32%) 155057173 bytes (33%) 159759339 bytes (34%) 164461547 bytes (35%) 169163755 bytes (36%) 173865963 bytes (37%) 178551787 bytes (38%) 183253995 bytes (39%) 187956203 bytes (40%) 192658411 bytes (41%) 197344235 bytes (42%) 202047488 bytes (43%) 206747763 bytes (44%) 211449215 bytes (45%) 216148006 bytes (46%) 220838032 bytes (47%) 225541366 bytes (48%) 230239897 bytes (49%) 234942105 bytes (50%) 239633511 bytes (51%) 244334592 bytes (52%) 249036800 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206257 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380600320 bytes (81%) 385302528 bytes (82%) 389988352 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887412 bytes (90%) 427589620 bytes (91%) 432275444 bytes (92%) 436977652 bytes (93%) 441679860 bytes (94%) 446382068 bytes (95%) 451084276 bytes (96%) 455770100 bytes (97%) 460472308 bytes (98%) 465176492 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("((x, m) => x':(x, m)) => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), segment = Embed(768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm((768,)), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768, 768, tanh)
      masklm => (
        transform => Chain(Dense(768, 768, gelu), LayerNorm((768,)))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768, 2), logsoftmax)
    )
)TransformerModel{Gpt{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("x':x => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().