Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419515 bytes (1%) 8826811 bytes (2%) 13234107 bytes (3%) 17641403 bytes (4%) 22065083 bytes (5%) 26472379 bytes (6%) 30879675 bytes (7%) 35286971 bytes (8%) 39694267 bytes (9%) 44117947 bytes (10%) 48525243 bytes (11%) 52932539 bytes (12%) 57339835 bytes (13%) 61747131 bytes (14%) 66154427 bytes (15%) 70578107 bytes (16%) 74985403 bytes (17%) 79392699 bytes (18%) 83799995 bytes (19%) 88207291 bytes (20%) 92630971 bytes (21%) 97038267 bytes (22%) 101445563 bytes (23%) 105852859 bytes (24%) 110260155 bytes (25%) 114667451 bytes (26%) 119091131 bytes (27%) 123498427 bytes (28%) 127905723 bytes (29%) 132313019 bytes (30%) 136720315 bytes (31%) 141143995 bytes (32%) 145551291 bytes (33%) 149958587 bytes (34%) 154365883 bytes (35%) 158773179 bytes (36%) 163180475 bytes (37%) 167604155 bytes (38%) 172011451 bytes (39%) 176418747 bytes (40%) 180826043 bytes (41%) 185233339 bytes (42%) 189657019 bytes (43%) 194064315 bytes (44%) 198471611 bytes (45%) 202878907 bytes (46%) 207286203 bytes (47%) 211693499 bytes (48%) 216117179 bytes (49%) 220524475 bytes (50%) 224931771 bytes (51%) 229339067 bytes (52%) 233746363 bytes (53%) 238170043 bytes (54%) 242577339 bytes (55%) 246984635 bytes (56%) 251391931 bytes (57%) 255799227 bytes (58%) 260206523 bytes (59%) 264630203 bytes (60%) 269037499 bytes (61%) 273444795 bytes (62%) 277852091 bytes (63%) 282259387 bytes (64%) 286683067 bytes (65%) 291090363 bytes (66%) 295497659 bytes (67%) 299904955 bytes (68%) 304312251 bytes (69%) 308719547 bytes (70%) 313143227 bytes (71%) 317550523 bytes (72%) 321957819 bytes (73%) 326365115 bytes (74%) 330772411 bytes (75%) 335196091 bytes (76%) 339603387 bytes (77%) 344010683 bytes (78%) 348417979 bytes (79%) 352825275 bytes (80%) 357232571 bytes (81%) 361656251 bytes (82%) 366063547 bytes (83%) 370470843 bytes (84%) 374878139 bytes (85%) 379285435 bytes (86%) 383709115 bytes (87%) 388116411 bytes (88%) 392523707 bytes (89%) 396931003 bytes (90%) 401338299 bytes (91%) 405745595 bytes (92%) 410169275 bytes (93%) 414576571 bytes (94%) 418983867 bytes (95%) 423391163 bytes (96%) 427798459 bytes (97%) 432222139 bytes (98%) 436629435 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714460 bytes (1%) 9400284 bytes (2%) 14102492 bytes (3%) 18804700 bytes (4%) 23506908 bytes (5%) 28192732 bytes (6%) 32894940 bytes (7%) 37597148 bytes (8%) 42299356 bytes (9%) 47001564 bytes (10%) 51687388 bytes (11%) 56389596 bytes (12%) 61091804 bytes (13%) 65794012 bytes (14%) 70479836 bytes (15%) 75182044 bytes (16%) 79884252 bytes (17%) 84586460 bytes (18%) 89288668 bytes (19%) 93974492 bytes (20%) 98676700 bytes (21%) 103378908 bytes (22%) 108081116 bytes (23%) 112783324 bytes (24%) 117469148 bytes (25%) 122171356 bytes (26%) 126873564 bytes (27%) 131575772 bytes (28%) 136261596 bytes (29%) 140963804 bytes (30%) 145666012 bytes (31%) 150368220 bytes (32%) 155070428 bytes (33%) 159756252 bytes (34%) 164458460 bytes (35%) 169160668 bytes (36%) 173862876 bytes (37%) 178548700 bytes (38%) 183250908 bytes (39%) 187953116 bytes (40%) 192655324 bytes (41%) 197357532 bytes (42%) 202043356 bytes (43%) 206745564 bytes (44%) 211447772 bytes (45%) 216149980 bytes (46%) 220852188 bytes (47%) 225538012 bytes (48%) 230240220 bytes (49%) 234942428 bytes (50%) 239644636 bytes (51%) 244330460 bytes (52%) 249032668 bytes (53%) 253734876 bytes (54%) 258437084 bytes (55%) 263139292 bytes (56%) 267825116 bytes (57%) 272527324 bytes (58%) 277229532 bytes (59%) 281931740 bytes (60%) 286617564 bytes (61%) 291319772 bytes (62%) 296021980 bytes (63%) 300724188 bytes (64%) 305426396 bytes (65%) 310112220 bytes (66%) 314814428 bytes (67%) 319516636 bytes (68%) 324218844 bytes (69%) 328921052 bytes (70%) 333606876 bytes (71%) 338309084 bytes (72%) 343011292 bytes (73%) 347713500 bytes (74%) 352399324 bytes (75%) 357101532 bytes (76%) 361803740 bytes (77%) 366505948 bytes (78%) 371208156 bytes (79%) 375893980 bytes (80%) 380596188 bytes (81%) 385298396 bytes (82%) 390000604 bytes (83%) 394686428 bytes (84%) 399388636 bytes (85%) 404090844 bytes (86%) 408793052 bytes (87%) 413495260 bytes (88%) 418181084 bytes (89%) 422883292 bytes (90%) 427585500 bytes (91%) 432287708 bytes (92%) 436973532 bytes (93%) 441675740 bytes (94%) 446377948 bytes (95%) 451080156 bytes (96%) 455782364 bytes (97%) 460468188 bytes (98%) 465170396 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().