Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419521 bytes (1%) 8826817 bytes (2%) 13234113 bytes (3%) 17641409 bytes (4%) 22065089 bytes (5%) 26472385 bytes (6%) 30879681 bytes (7%) 35286977 bytes (8%) 39694273 bytes (9%) 44117953 bytes (10%) 48525249 bytes (11%) 52932545 bytes (12%) 57339841 bytes (13%) 61747137 bytes (14%) 66154433 bytes (15%) 70578113 bytes (16%) 74985409 bytes (17%) 79392705 bytes (18%) 83800001 bytes (19%) 88207297 bytes (20%) 92630977 bytes (21%) 97038273 bytes (22%) 101445569 bytes (23%) 105852865 bytes (24%) 110260161 bytes (25%) 114667457 bytes (26%) 119091137 bytes (27%) 123498433 bytes (28%) 127905729 bytes (29%) 132313025 bytes (30%) 136720321 bytes (31%) 141144001 bytes (32%) 145551297 bytes (33%) 149958593 bytes (34%) 154365889 bytes (35%) 158773185 bytes (36%) 163180481 bytes (37%) 167604161 bytes (38%) 172011457 bytes (39%) 176418753 bytes (40%) 180826049 bytes (41%) 185233345 bytes (42%) 189657025 bytes (43%) 194064321 bytes (44%) 198471617 bytes (45%) 202878913 bytes (46%) 207286209 bytes (47%) 211693505 bytes (48%) 216117185 bytes (49%) 220524481 bytes (50%) 224931777 bytes (51%) 229339073 bytes (52%) 233746369 bytes (53%) 238170049 bytes (54%) 242577345 bytes (55%) 246984641 bytes (56%) 251391937 bytes (57%) 255799233 bytes (58%) 260206529 bytes (59%) 264630209 bytes (60%) 269037505 bytes (61%) 273444801 bytes (62%) 277852097 bytes (63%) 282259393 bytes (64%) 286683073 bytes (65%) 291090369 bytes (66%) 295497665 bytes (67%) 299904961 bytes (68%) 304312257 bytes (69%) 308719553 bytes (70%) 313143233 bytes (71%) 317550529 bytes (72%) 321957825 bytes (73%) 326365121 bytes (74%) 330772417 bytes (75%) 335196097 bytes (76%) 339603393 bytes (77%) 344010689 bytes (78%) 348417985 bytes (79%) 352825281 bytes (80%) 357232577 bytes (81%) 361656257 bytes (82%) 366063553 bytes (83%) 370470849 bytes (84%) 374878145 bytes (85%) 379285441 bytes (86%) 383709121 bytes (87%) 388116417 bytes (88%) 392523713 bytes (89%) 396931009 bytes (90%) 401338305 bytes (91%) 405745601 bytes (92%) 410169281 bytes (93%) 414576577 bytes (94%) 418983873 bytes (95%) 423391169 bytes (96%) 427798465 bytes (97%) 432222145 bytes (98%) 436629441 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14097154 bytes (3%) 18799362 bytes (4%) 23501570 bytes (5%) 28203778 bytes (6%) 32905986 bytes (7%) 37591810 bytes (8%) 42294018 bytes (9%) 46996226 bytes (10%) 51698434 bytes (11%) 56384258 bytes (12%) 61086466 bytes (13%) 65788674 bytes (14%) 70490882 bytes (15%) 75193090 bytes (16%) 79878914 bytes (17%) 84581122 bytes (18%) 89283330 bytes (19%) 93985538 bytes (20%) 98687746 bytes (21%) 103373570 bytes (22%) 108075778 bytes (23%) 112777986 bytes (24%) 117480194 bytes (25%) 122166018 bytes (26%) 126868226 bytes (27%) 131570434 bytes (28%) 136272642 bytes (29%) 140974850 bytes (30%) 145660674 bytes (31%) 150362882 bytes (32%) 155065090 bytes (33%) 159767298 bytes (34%) 164453122 bytes (35%) 169155330 bytes (36%) 173857538 bytes (37%) 178559746 bytes (38%) 183261954 bytes (39%) 187947778 bytes (40%) 192649986 bytes (41%) 197352194 bytes (42%) 202054402 bytes (43%) 206740226 bytes (44%) 211442434 bytes (45%) 216144642 bytes (46%) 220846850 bytes (47%) 225549058 bytes (48%) 230234882 bytes (49%) 234937090 bytes (50%) 239639298 bytes (51%) 244341506 bytes (52%) 249043714 bytes (53%) 253729538 bytes (54%) 258431746 bytes (55%) 263133954 bytes (56%) 267836162 bytes (57%) 272521986 bytes (58%) 277224194 bytes (59%) 281926402 bytes (60%) 286628610 bytes (61%) 291330818 bytes (62%) 296016642 bytes (63%) 300718850 bytes (64%) 305421058 bytes (65%) 310123266 bytes (66%) 314809090 bytes (67%) 319511298 bytes (68%) 324213506 bytes (69%) 328915714 bytes (70%) 333617922 bytes (71%) 338303746 bytes (72%) 343005954 bytes (73%) 347708162 bytes (74%) 352410370 bytes (75%) 357112578 bytes (76%) 361798402 bytes (77%) 366500610 bytes (78%) 371202818 bytes (79%) 375905026 bytes (80%) 380590850 bytes (81%) 385300085 bytes (82%) 390002293 bytes (83%) 394688117 bytes (84%) 399390325 bytes (85%) 404092533 bytes (86%) 408794741 bytes (87%) 413480565 bytes (88%) 418182773 bytes (89%) 422884981 bytes (90%) 427587189 bytes (91%) 432289397 bytes (92%) 436975221 bytes (93%) 441677429 bytes (94%) 446379637 bytes (95%) 451081845 bytes (96%) 455784053 bytes (97%) 460477493 bytes (98%) 465179701 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().