Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419461 bytes (1%) 8826757 bytes (2%) 13234053 bytes (3%) 17641349 bytes (4%) 22065029 bytes (5%) 26472325 bytes (6%) 30879621 bytes (7%) 35286917 bytes (8%) 39694213 bytes (9%) 44117893 bytes (10%) 48525189 bytes (11%) 52932485 bytes (12%) 57339781 bytes (13%) 61747077 bytes (14%) 66154373 bytes (15%) 70578053 bytes (16%) 74985349 bytes (17%) 79392645 bytes (18%) 83799941 bytes (19%) 88207237 bytes (20%) 92630917 bytes (21%) 97038213 bytes (22%) 101445509 bytes (23%) 105852805 bytes (24%) 110260101 bytes (25%) 114667397 bytes (26%) 119091077 bytes (27%) 123498373 bytes (28%) 127905669 bytes (29%) 132312965 bytes (30%) 136720261 bytes (31%) 141143941 bytes (32%) 145551237 bytes (33%) 149958533 bytes (34%) 154365829 bytes (35%) 158773125 bytes (36%) 163180421 bytes (37%) 167604101 bytes (38%) 172011397 bytes (39%) 176418693 bytes (40%) 180825989 bytes (41%) 185233285 bytes (42%) 189656965 bytes (43%) 194064261 bytes (44%) 198471557 bytes (45%) 202878853 bytes (46%) 207286149 bytes (47%) 211693445 bytes (48%) 216117125 bytes (49%) 220524421 bytes (50%) 224931717 bytes (51%) 229339013 bytes (52%) 233746309 bytes (53%) 238169989 bytes (54%) 242577285 bytes (55%) 246984581 bytes (56%) 251391877 bytes (57%) 255799173 bytes (58%) 260206469 bytes (59%) 264630149 bytes (60%) 269037445 bytes (61%) 273444741 bytes (62%) 277852037 bytes (63%) 282259333 bytes (64%) 286683013 bytes (65%) 291090309 bytes (66%) 295497605 bytes (67%) 299904901 bytes (68%) 304312197 bytes (69%) 308719493 bytes (70%) 313143173 bytes (71%) 317550469 bytes (72%) 321957765 bytes (73%) 326365061 bytes (74%) 330772357 bytes (75%) 335196037 bytes (76%) 339603333 bytes (77%) 344010629 bytes (78%) 348417925 bytes (79%) 352825221 bytes (80%) 357232517 bytes (81%) 361656197 bytes (82%) 366063493 bytes (83%) 370470789 bytes (84%) 374878085 bytes (85%) 379285381 bytes (86%) 383709061 bytes (87%) 388116357 bytes (88%) 392523653 bytes (89%) 396930949 bytes (90%) 401338245 bytes (91%) 405745541 bytes (92%) 410169221 bytes (93%) 414576517 bytes (94%) 418983813 bytes (95%) 423391109 bytes (96%) 427798405 bytes (97%) 432222085 bytes (98%) 436629381 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714405 bytes (1%) 9400229 bytes (2%) 14102437 bytes (3%) 18804645 bytes (4%) 23506853 bytes (5%) 28192677 bytes (6%) 32894885 bytes (7%) 37597093 bytes (8%) 42299301 bytes (9%) 47001509 bytes (10%) 51687333 bytes (11%) 56389541 bytes (12%) 61091749 bytes (13%) 65793957 bytes (14%) 70479781 bytes (15%) 75181989 bytes (16%) 79884197 bytes (17%) 84586405 bytes (18%) 89288613 bytes (19%) 93974437 bytes (20%) 98676645 bytes (21%) 103378853 bytes (22%) 108081061 bytes (23%) 112783269 bytes (24%) 117469093 bytes (25%) 122171301 bytes (26%) 126873509 bytes (27%) 131575717 bytes (28%) 136261541 bytes (29%) 140963749 bytes (30%) 145665957 bytes (31%) 150368165 bytes (32%) 155070373 bytes (33%) 159756197 bytes (34%) 164458405 bytes (35%) 169160613 bytes (36%) 173862821 bytes (37%) 178548645 bytes (38%) 183250853 bytes (39%) 187953061 bytes (40%) 192655269 bytes (41%) 197357477 bytes (42%) 202043301 bytes (43%) 206745509 bytes (44%) 211447717 bytes (45%) 216149925 bytes (46%) 220852133 bytes (47%) 225537957 bytes (48%) 230240165 bytes (49%) 234942373 bytes (50%) 239644581 bytes (51%) 244330405 bytes (52%) 249032613 bytes (53%) 253734821 bytes (54%) 258437029 bytes (55%) 263139237 bytes (56%) 267825061 bytes (57%) 272527269 bytes (58%) 277229477 bytes (59%) 281931685 bytes (60%) 286617509 bytes (61%) 291319717 bytes (62%) 296021925 bytes (63%) 300724133 bytes (64%) 305426341 bytes (65%) 310112165 bytes (66%) 314814373 bytes (67%) 319516581 bytes (68%) 324218789 bytes (69%) 328920997 bytes (70%) 333606821 bytes (71%) 338309029 bytes (72%) 343011237 bytes (73%) 347713445 bytes (74%) 352399269 bytes (75%) 357101477 bytes (76%) 361803685 bytes (77%) 366505893 bytes (78%) 371208101 bytes (79%) 375893925 bytes (80%) 380596133 bytes (81%) 385298341 bytes (82%) 390000549 bytes (83%) 394686373 bytes (84%) 399388581 bytes (85%) 404090789 bytes (86%) 408792997 bytes (87%) 413495205 bytes (88%) 418181029 bytes (89%) 422883237 bytes (90%) 427585445 bytes (91%) 432287653 bytes (92%) 436989861 bytes (93%) 441675685 bytes (94%) 446377893 bytes (95%) 451080101 bytes (96%) 455782309 bytes (97%) 460468133 bytes (98%) 465170341 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().