Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419136 bytes (1%) 8830976 bytes (2%) 13238272 bytes (3%) 17645568 bytes (4%) 22052864 bytes (5%) 26476544 bytes (6%) 30883840 bytes (7%) 35291136 bytes (8%) 39698432 bytes (9%) 44104944 bytes (10%) 48513024 bytes (11%) 52936704 bytes (12%) 57344000 bytes (13%) 61751296 bytes (14%) 66157725 bytes (15%) 70565021 bytes (16%) 74989568 bytes (17%) 79395925 bytes (18%) 83796701 bytes (19%) 88220381 bytes (20%) 92627677 bytes (21%) 97026720 bytes (22%) 101450400 bytes (23%) 105862708 bytes (24%) 110269316 bytes (25%) 114675633 bytes (26%) 119081921 bytes (27%) 123489217 bytes (28%) 127912113 bytes (29%) 132308076 bytes (30%) 136731756 bytes (31%) 141138302 bytes (32%) 145545598 bytes (33%) 149952788 bytes (34%) 154360084 bytes (35%) 158783764 bytes (36%) 163190610 bytes (37%) 167596990 bytes (38%) 172004286 bytes (39%) 176411582 bytes (40%) 180835872 bytes (41%) 185242287 bytes (42%) 189648594 bytes (43%) 194065084 bytes (44%) 198464152 bytes (45%) 202887832 bytes (46%) 207298611 bytes (47%) 211705907 bytes (48%) 216113203 bytes (49%) 220520499 bytes (50%) 224927074 bytes (51%) 229342529 bytes (52%) 233749825 bytes (53%) 238155776 bytes (54%) 242579456 bytes (55%) 246986001 bytes (56%) 251392576 bytes (57%) 255799073 bytes (58%) 260208298 bytes (59%) 264631978 bytes (60%) 269038269 bytes (61%) 273445253 bytes (62%) 277852549 bytes (63%) 282258943 bytes (64%) 286682623 bytes (65%) 291089705 bytes (66%) 295502977 bytes (67%) 299899606 bytes (68%) 304322262 bytes (69%) 308727591 bytes (70%) 313134887 bytes (71%) 317541194 bytes (72%) 321960696 bytes (73%) 326367992 bytes (74%) 330775288 bytes (75%) 335181559 bytes (76%) 339605239 bytes (77%) 344012535 bytes (78%) 348418048 bytes (79%) 352828416 bytes (80%) 357232680 bytes (81%) 361656360 bytes (82%) 366063308 bytes (83%) 370470181 bytes (84%) 374877477 bytes (85%) 379284082 bytes (86%) 383707762 bytes (87%) 388114362 bytes (88%) 392521658 bytes (89%) 396928954 bytes (90%) 401336250 bytes (91%) 405759930 bytes (92%) 410166495 bytes (93%) 414578559 bytes (94%) 418985129 bytes (95%) 423393151 bytes (96%) 427799446 bytes (97%) 432222154 bytes (98%) 436629450 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14106624 bytes (3%) 18808832 bytes (4%) 23494656 bytes (5%) 28196864 bytes (6%) 32899072 bytes (7%) 37600935 bytes (8%) 42303143 bytes (9%) 46988967 bytes (10%) 51691134 bytes (11%) 56393342 bytes (12%) 61095550 bytes (13%) 65781374 bytes (14%) 70483582 bytes (15%) 75185790 bytes (16%) 79887998 bytes (17%) 84590206 bytes (18%) 89276030 bytes (19%) 93978238 bytes (20%) 98680446 bytes (21%) 103382654 bytes (22%) 108084588 bytes (23%) 112776758 bytes (24%) 117478966 bytes (25%) 122180608 bytes (26%) 126865663 bytes (27%) 131567616 bytes (28%) 136269824 bytes (29%) 140972021 bytes (30%) 145657845 bytes (31%) 150360053 bytes (32%) 155062261 bytes (33%) 159764469 bytes (34%) 164466677 bytes (35%) 169152501 bytes (36%) 173854709 bytes (37%) 178551161 bytes (38%) 183253150 bytes (39%) 187955358 bytes (40%) 192657566 bytes (41%) 197343390 bytes (42%) 202045598 bytes (43%) 206747806 bytes (44%) 211440126 bytes (45%) 216146571 bytes (46%) 220848131 bytes (47%) 225535978 bytes (48%) 230237808 bytes (49%) 234940016 bytes (50%) 239631124 bytes (51%) 244333332 bytes (52%) 249043033 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206592 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195447 bytes (79%) 375897655 bytes (80%) 380599863 bytes (81%) 385301402 bytes (82%) 389987226 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887424 bytes (90%) 427589632 bytes (91%) 432275456 bytes (92%) 436977664 bytes (93%) 441679872 bytes (94%) 446382080 bytes (95%) 451084288 bytes (96%) 455770112 bytes (97%) 460472320 bytes (98%) 465174528 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("((x, m) => x':(x, m)) => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), segment = Embed(768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm((768,)), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768, 768, tanh)
      masklm => (
        transform => Chain(Dense(768, 768, gelu), LayerNorm((768,)))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768, 2), logsoftmax)
    )
)TransformerModel{Gpt{Stack{NTuple{12, Transformer{MultiheadAttention{Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}, Flux.Dropout{Float64, Colon}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Transformers.Basic.PwFFN{Flux.Dense{typeof(NNlib.gelu), Matrix{Float32}, Vector{Float32}}, Flux.Dense{typeof(identity), Matrix{Float32}, Vector{Float32}}}, Flux.LayerNorm{typeof(identity), Flux.Diagonal{Vector{Float32}}, Float32, 1}, Flux.Dropout{Float64, Colon}}}, Symbol("x':x => 12")}, Flux.Dropout{Float64, Colon}}}(
  embed = CompositeEmbedding(tok = Embed(768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().