Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419555 bytes (1%) 8826851 bytes (2%) 13234147 bytes (3%) 17641443 bytes (4%) 22065123 bytes (5%) 26472419 bytes (6%) 30879715 bytes (7%) 35287011 bytes (8%) 39694307 bytes (9%) 44117987 bytes (10%) 48525283 bytes (11%) 52932579 bytes (12%) 57339875 bytes (13%) 61747171 bytes (14%) 66154467 bytes (15%) 70578147 bytes (16%) 74985443 bytes (17%) 79392739 bytes (18%) 83800035 bytes (19%) 88207331 bytes (20%) 92631011 bytes (21%) 97038307 bytes (22%) 101445603 bytes (23%) 105852899 bytes (24%) 110260195 bytes (25%) 114667491 bytes (26%) 119091171 bytes (27%) 123498467 bytes (28%) 127905763 bytes (29%) 132313059 bytes (30%) 136720355 bytes (31%) 141144035 bytes (32%) 145551331 bytes (33%) 149958627 bytes (34%) 154365923 bytes (35%) 158773219 bytes (36%) 163180515 bytes (37%) 167604195 bytes (38%) 172011491 bytes (39%) 176418787 bytes (40%) 180826083 bytes (41%) 185233379 bytes (42%) 189657059 bytes (43%) 194064355 bytes (44%) 198471651 bytes (45%) 202878947 bytes (46%) 207286243 bytes (47%) 211693539 bytes (48%) 216117219 bytes (49%) 220524515 bytes (50%) 224931811 bytes (51%) 229339107 bytes (52%) 233746403 bytes (53%) 238170083 bytes (54%) 242577379 bytes (55%) 246984675 bytes (56%) 251391971 bytes (57%) 255799267 bytes (58%) 260206563 bytes (59%) 264630243 bytes (60%) 269037539 bytes (61%) 273444835 bytes (62%) 277852131 bytes (63%) 282259427 bytes (64%) 286683107 bytes (65%) 291090403 bytes (66%) 295497699 bytes (67%) 299904995 bytes (68%) 304312291 bytes (69%) 308719587 bytes (70%) 313143267 bytes (71%) 317550563 bytes (72%) 321957859 bytes (73%) 326365155 bytes (74%) 330772451 bytes (75%) 335196131 bytes (76%) 339603427 bytes (77%) 344010723 bytes (78%) 348418019 bytes (79%) 352825315 bytes (80%) 357232611 bytes (81%) 361656291 bytes (82%) 366063587 bytes (83%) 370470883 bytes (84%) 374878179 bytes (85%) 379285475 bytes (86%) 383709155 bytes (87%) 388116451 bytes (88%) 392523747 bytes (89%) 396931043 bytes (90%) 401338339 bytes (91%) 405745635 bytes (92%) 410169315 bytes (93%) 414576611 bytes (94%) 418983907 bytes (95%) 423391203 bytes (96%) 427798499 bytes (97%) 432222179 bytes (98%) 436629475 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714485 bytes (1%) 9400309 bytes (2%) 14102517 bytes (3%) 18804725 bytes (4%) 23506933 bytes (5%) 28192757 bytes (6%) 32894965 bytes (7%) 37597173 bytes (8%) 42299381 bytes (9%) 47001589 bytes (10%) 51687413 bytes (11%) 56389621 bytes (12%) 61091829 bytes (13%) 65794037 bytes (14%) 70479861 bytes (15%) 75182069 bytes (16%) 79884277 bytes (17%) 84586485 bytes (18%) 89288693 bytes (19%) 93974517 bytes (20%) 98676725 bytes (21%) 103378933 bytes (22%) 108081141 bytes (23%) 112783349 bytes (24%) 117469173 bytes (25%) 122171381 bytes (26%) 126873589 bytes (27%) 131575797 bytes (28%) 136261621 bytes (29%) 140963829 bytes (30%) 145666037 bytes (31%) 150368245 bytes (32%) 155070453 bytes (33%) 159756277 bytes (34%) 164458485 bytes (35%) 169160693 bytes (36%) 173862901 bytes (37%) 178548725 bytes (38%) 183250933 bytes (39%) 187953141 bytes (40%) 192655349 bytes (41%) 197357557 bytes (42%) 202043381 bytes (43%) 206745589 bytes (44%) 211447797 bytes (45%) 216150005 bytes (46%) 220852213 bytes (47%) 225538037 bytes (48%) 230240245 bytes (49%) 234942453 bytes (50%) 239644661 bytes (51%) 244330485 bytes (52%) 249032693 bytes (53%) 253734901 bytes (54%) 258437109 bytes (55%) 263139317 bytes (56%) 267825141 bytes (57%) 272527349 bytes (58%) 277229557 bytes (59%) 281931765 bytes (60%) 286617589 bytes (61%) 291319797 bytes (62%) 296022005 bytes (63%) 300724213 bytes (64%) 305426421 bytes (65%) 310112245 bytes (66%) 314814453 bytes (67%) 319516661 bytes (68%) 324218869 bytes (69%) 328921077 bytes (70%) 333606901 bytes (71%) 338309109 bytes (72%) 343011317 bytes (73%) 347713525 bytes (74%) 352399349 bytes (75%) 357101557 bytes (76%) 361803765 bytes (77%) 366505973 bytes (78%) 371208181 bytes (79%) 375894005 bytes (80%) 380596213 bytes (81%) 385298421 bytes (82%) 390000629 bytes (83%) 394686453 bytes (84%) 399388661 bytes (85%) 404090869 bytes (86%) 408793077 bytes (87%) 413495285 bytes (88%) 418181109 bytes (89%) 422883317 bytes (90%) 427585525 bytes (91%) 432287733 bytes (92%) 436973557 bytes (93%) 441675765 bytes (94%) 446377973 bytes (95%) 451080181 bytes (96%) 455782389 bytes (97%) 460468213 bytes (98%) 465170421 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().