Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4419516 bytes (1%) 8826812 bytes (2%) 13234108 bytes (3%) 17641404 bytes (4%) 22065084 bytes (5%) 26472380 bytes (6%) 30879676 bytes (7%) 35286972 bytes (8%) 39694268 bytes (9%) 44117948 bytes (10%) 48525244 bytes (11%) 52932540 bytes (12%) 57339836 bytes (13%) 61747132 bytes (14%) 66154428 bytes (15%) 70578108 bytes (16%) 74985404 bytes (17%) 79392700 bytes (18%) 83799996 bytes (19%) 88207292 bytes (20%) 92630972 bytes (21%) 97038268 bytes (22%) 101445564 bytes (23%) 105852860 bytes (24%) 110260156 bytes (25%) 114667452 bytes (26%) 119091132 bytes (27%) 123498428 bytes (28%) 127905724 bytes (29%) 132313020 bytes (30%) 136720316 bytes (31%) 141143996 bytes (32%) 145551292 bytes (33%) 149958588 bytes (34%) 154365884 bytes (35%) 158773180 bytes (36%) 163180476 bytes (37%) 167604156 bytes (38%) 172011452 bytes (39%) 176418748 bytes (40%) 180826044 bytes (41%) 185233340 bytes (42%) 189657020 bytes (43%) 194064316 bytes (44%) 198471612 bytes (45%) 202878908 bytes (46%) 207286204 bytes (47%) 211693500 bytes (48%) 216117180 bytes (49%) 220524476 bytes (50%) 224931772 bytes (51%) 229339068 bytes (52%) 233746364 bytes (53%) 238170044 bytes (54%) 242577340 bytes (55%) 246984636 bytes (56%) 251391932 bytes (57%) 255799228 bytes (58%) 260206524 bytes (59%) 264630204 bytes (60%) 269037500 bytes (61%) 273444796 bytes (62%) 277852092 bytes (63%) 282259388 bytes (64%) 286683068 bytes (65%) 291090364 bytes (66%) 295497660 bytes (67%) 299904956 bytes (68%) 304312252 bytes (69%) 308719548 bytes (70%) 313143228 bytes (71%) 317550524 bytes (72%) 321957820 bytes (73%) 326365116 bytes (74%) 330772412 bytes (75%) 335196092 bytes (76%) 339603388 bytes (77%) 344010684 bytes (78%) 348417980 bytes (79%) 352825276 bytes (80%) 357232572 bytes (81%) 361656252 bytes (82%) 366063548 bytes (83%) 370470844 bytes (84%) 374878140 bytes (85%) 379285436 bytes (86%) 383709116 bytes (87%) 388116412 bytes (88%) 392523708 bytes (89%) 396931004 bytes (90%) 401338300 bytes (91%) 405745596 bytes (92%) 410169276 bytes (93%) 414576572 bytes (94%) 418983868 bytes (95%) 423391164 bytes (96%) 427798460 bytes (97%) 432222140 bytes (98%) 436629436 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714460 bytes (1%) 9400284 bytes (2%) 14102492 bytes (3%) 18804700 bytes (4%) 23506908 bytes (5%) 28192732 bytes (6%) 32894940 bytes (7%) 37597148 bytes (8%) 42299356 bytes (9%) 47001564 bytes (10%) 51687388 bytes (11%) 56389596 bytes (12%) 61091804 bytes (13%) 65794012 bytes (14%) 70479836 bytes (15%) 75182044 bytes (16%) 79884252 bytes (17%) 84586460 bytes (18%) 89288668 bytes (19%) 93974492 bytes (20%) 98676700 bytes (21%) 103378908 bytes (22%) 108081116 bytes (23%) 112783324 bytes (24%) 117469148 bytes (25%) 122171356 bytes (26%) 126873564 bytes (27%) 131575772 bytes (28%) 136261596 bytes (29%) 140963804 bytes (30%) 145670144 bytes (31%) 150372352 bytes (32%) 155058176 bytes (33%) 159760384 bytes (34%) 164462592 bytes (35%) 169164800 bytes (36%) 173850624 bytes (37%) 178552832 bytes (38%) 183255040 bytes (39%) 187957248 bytes (40%) 192659456 bytes (41%) 197345280 bytes (42%) 202053940 bytes (43%) 206756148 bytes (44%) 211441972 bytes (45%) 216144180 bytes (46%) 220846388 bytes (47%) 225548596 bytes (48%) 230234420 bytes (49%) 234936628 bytes (50%) 239638836 bytes (51%) 244341044 bytes (52%) 249043252 bytes (53%) 253729076 bytes (54%) 258431284 bytes (55%) 263133492 bytes (56%) 267835700 bytes (57%) 272521524 bytes (58%) 277223732 bytes (59%) 281925940 bytes (60%) 286628148 bytes (61%) 291330356 bytes (62%) 296024530 bytes (63%) 300726738 bytes (64%) 305412562 bytes (65%) 310114770 bytes (66%) 314816978 bytes (67%) 319519186 bytes (68%) 324221394 bytes (69%) 328909500 bytes (70%) 333611708 bytes (71%) 338309988 bytes (72%) 343012196 bytes (73%) 347714404 bytes (74%) 352400228 bytes (75%) 357102436 bytes (76%) 361804644 bytes (77%) 366506852 bytes (78%) 371192676 bytes (79%) 375894884 bytes (80%) 380597092 bytes (81%) 385299300 bytes (82%) 390001508 bytes (83%) 394687332 bytes (84%) 399389540 bytes (85%) 404091748 bytes (86%) 408793956 bytes (87%) 413496164 bytes (88%) 418181988 bytes (89%) 422884196 bytes (90%) 427586404 bytes (91%) 432288612 bytes (92%) 436974436 bytes (93%) 441676644 bytes (94%) 446378852 bytes (95%) 451081060 bytes (96%) 455783268 bytes (97%) 460469092 bytes (98%) 465171300 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().