Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4423680 bytes (1%) 8830976 bytes (2%) 13238272 bytes (3%) 17645568 bytes (4%) 22052864 bytes (5%) 26476544 bytes (6%) 30883840 bytes (7%) 35291136 bytes (8%) 39698432 bytes (9%) 44105728 bytes (10%) 48513024 bytes (11%) 52936704 bytes (12%) 57344000 bytes (13%) 61751296 bytes (14%) 66158592 bytes (15%) 70565888 bytes (16%) 74989568 bytes (17%) 79396864 bytes (18%) 83804160 bytes (19%) 88211456 bytes (20%) 92618752 bytes (21%) 97026048 bytes (22%) 101449728 bytes (23%) 105857024 bytes (24%) 110264320 bytes (25%) 114671616 bytes (26%) 119078912 bytes (27%) 123502592 bytes (28%) 127909888 bytes (29%) 132317184 bytes (30%) 136724480 bytes (31%) 141131776 bytes (32%) 145539072 bytes (33%) 149962752 bytes (34%) 154370048 bytes (35%) 158777344 bytes (36%) 163184640 bytes (37%) 167591936 bytes (38%) 172015616 bytes (39%) 176422912 bytes (40%) 180830208 bytes (41%) 185237504 bytes (42%) 189644800 bytes (43%) 194052096 bytes (44%) 198475776 bytes (45%) 202883072 bytes (46%) 207290368 bytes (47%) 211697664 bytes (48%) 216104960 bytes (49%) 220528640 bytes (50%) 224935936 bytes (51%) 229343232 bytes (52%) 233750528 bytes (53%) 238157824 bytes (54%) 242565120 bytes (55%) 246988800 bytes (56%) 251396096 bytes (57%) 255803392 bytes (58%) 260210688 bytes (59%) 264617984 bytes (60%) 269041664 bytes (61%) 273448960 bytes (62%) 277856256 bytes (63%) 282263552 bytes (64%) 286670848 bytes (65%) 291078144 bytes (66%) 295501824 bytes (67%) 299909120 bytes (68%) 304316416 bytes (69%) 308723712 bytes (70%) 313131008 bytes (71%) 317554688 bytes (72%) 321961984 bytes (73%) 326369280 bytes (74%) 330776576 bytes (75%) 335183872 bytes (76%) 339591168 bytes (77%) 344014848 bytes (78%) 348422144 bytes (79%) 352829440 bytes (80%) 357236736 bytes (81%) 361644032 bytes (82%) 366067712 bytes (83%) 370475008 bytes (84%) 374882304 bytes (85%) 379289600 bytes (86%) 383696896 bytes (87%) 388104192 bytes (88%) 392527872 bytes (89%) 396935168 bytes (90%) 401342464 bytes (91%) 405749760 bytes (92%) 410157056 bytes (93%) 414580736 bytes (94%) 418988032 bytes (95%) 423395328 bytes (96%) 427802624 bytes (97%) 432209920 bytes (98%) 436617216 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4714399 bytes (1%) 9400223 bytes (2%) 14102431 bytes (3%) 18804639 bytes (4%) 23506847 bytes (5%) 28192671 bytes (6%) 32894879 bytes (7%) 37597087 bytes (8%) 42299295 bytes (9%) 47001503 bytes (10%) 51687327 bytes (11%) 56389535 bytes (12%) 61091743 bytes (13%) 65793951 bytes (14%) 70479775 bytes (15%) 75181983 bytes (16%) 79884191 bytes (17%) 84586399 bytes (18%) 89288607 bytes (19%) 93974431 bytes (20%) 98676639 bytes (21%) 103378847 bytes (22%) 108081055 bytes (23%) 112783263 bytes (24%) 117469087 bytes (25%) 122171295 bytes (26%) 126873503 bytes (27%) 131575711 bytes (28%) 136261535 bytes (29%) 140963743 bytes (30%) 145665951 bytes (31%) 150368159 bytes (32%) 155070367 bytes (33%) 159756191 bytes (34%) 164458399 bytes (35%) 169160607 bytes (36%) 173862815 bytes (37%) 178548639 bytes (38%) 183250847 bytes (39%) 187953055 bytes (40%) 192655263 bytes (41%) 197357471 bytes (42%) 202043295 bytes (43%) 206745503 bytes (44%) 211447711 bytes (45%) 216149919 bytes (46%) 220852127 bytes (47%) 225537951 bytes (48%) 230240159 bytes (49%) 234942367 bytes (50%) 239644575 bytes (51%) 244330399 bytes (52%) 249032607 bytes (53%) 253734815 bytes (54%) 258437023 bytes (55%) 263139231 bytes (56%) 267825055 bytes (57%) 272527263 bytes (58%) 277229471 bytes (59%) 281931679 bytes (60%) 286617503 bytes (61%) 291319711 bytes (62%) 296021919 bytes (63%) 300724127 bytes (64%) 305426335 bytes (65%) 310112159 bytes (66%) 314814367 bytes (67%) 319516575 bytes (68%) 324218783 bytes (69%) 328920991 bytes (70%) 333606815 bytes (71%) 338309023 bytes (72%) 343011231 bytes (73%) 347713439 bytes (74%) 352399263 bytes (75%) 357101471 bytes (76%) 361803679 bytes (77%) 366505887 bytes (78%) 371208095 bytes (79%) 375893919 bytes (80%) 380596127 bytes (81%) 385298335 bytes (82%) 390000543 bytes (83%) 394686367 bytes (84%) 399388575 bytes (85%) 404090783 bytes (86%) 408792991 bytes (87%) 413495199 bytes (88%) 418181023 bytes (89%) 422883231 bytes (90%) 427585439 bytes (91%) 432287647 bytes (92%) 436989855 bytes (93%) 441675679 bytes (94%) 446377887 bytes (95%) 451080095 bytes (96%) 455782303 bytes (97%) 460468127 bytes (98%) 465170335 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().