Transformers.Pretrain

Functions for download and loading pretrain models.

using Pretrains

For GPT and BERT, we provide a simple api to get the released pretrain weight and load them into our Julia version Transformer implementation.

using Transformers
using Transformers.Pretrain
using Transformers.GenerativePreTrain
using Transformers.BidirectionalEncoder

# disable cli download check
ENV["DATADEPS_ALWAYS_ACCEPT"] = true

#load everything in the pretrain model
bert_model, wordpiece, tokenizer = pretrain"Bert-uncased_L-12_H-768_A-12"

#load model weight only
gpt_model = pretrain"gpt-OpenAIftlm:gpt_model"

#show the loaded model
show(bert_model)
show(gpt_model)

Total: 441026658 bytes
Downloaded: 4418905 bytes (1%) 8830512 bytes (2%) 13238152 bytes (3%) 17644655 bytes (4%) 22051906 bytes (5%) 26474804 bytes (6%) 30881677 bytes (7%) 35290443 bytes (8%) 39696946 bytes (9%) 44105339 bytes (10%) 48512979 bytes (11%) 52936236 bytes (12%) 57342731 bytes (13%) 61751131 bytes (14%) 66157252 bytes (15%) 70564892 bytes (16%) 74986630 bytes (17%) 79393125 bytes (18%) 83803800 bytes (19%) 88209918 bytes (20%) 92616795 bytes (21%) 97030130 bytes (22%) 101436611 bytes (23%) 105850577 bytes (24%) 110261248 bytes (25%) 114667520 bytes (26%) 119082404 bytes (27%) 123501494 bytes (28%) 127903626 bytes (29%) 132317031 bytes (30%) 136725310 bytes (31%) 141135073 bytes (32%) 145542324 bytes (33%) 149956309 bytes (34%) 154366976 bytes (35%) 158780866 bytes (36%) 163194832 bytes (37%) 167605156 bytes (38%) 172004233 bytes (39%) 176421724 bytes (40%) 180835116 bytes (41%) 185242382 bytes (42%) 189648052 bytes (43%) 194055168 bytes (44%) 198471680 bytes (45%) 202886579 bytes (46%) 207284161 bytes (47%) 211707797 bytes (48%) 216110966 bytes (49%) 220522314 bytes (50%) 224935713 bytes (51%) 229348095 bytes (52%) 233753765 bytes (53%) 238154752 bytes (54%) 242576384 bytes (55%) 246988767 bytes (56%) 251394998 bytes (57%) 255801236 bytes (58%) 260213607 bytes (59%) 264616777 bytes (60%) 269028128 bytes (61%) 273442558 bytes (62%) 277848411 bytes (63%) 282264562 bytes (64%) 286675968 bytes (65%) 291093461 bytes (66%) 295501746 bytes (67%) 299898757 bytes (68%) 304313177 bytes (69%) 308726590 bytes (70%) 313141010 bytes (71%) 317553431 bytes (72%) 321949696 bytes (73%) 326370304 bytes (74%) 330771442 bytes (75%) 335194050 bytes (76%) 339591079 bytes (77%) 344004471 bytes (78%) 348427082 bytes (79%) 352824108 bytes (80%) 357240018 bytes (81%) 361653988 bytes (82%) 366065664 bytes (83%) 370466802 bytes (84%) 374885811 bytes (85%) 379291585 bytes (86%) 383695768 bytes (87%) 388107579 bytes (88%) 392516425 bytes (89%) 396928797 bytes (90%) 401347330 bytes (91%) 405760722 bytes (92%) 410160128 bytes (93%) 414578688 bytes (94%) 418975697 bytes (95%) 423401398 bytes (96%) 427798406 bytes (97%) 432208743 bytes (98%) 436622139 bytes (99%) 441026658 bytes (100%)
[ Info: loading pretrain bert model: uncased_L-12_H-768_A-12.tfbson
Total: 469864010 bytes
Downloaded: 4702208 bytes (1%) 9404416 bytes (2%) 14106624 bytes (3%) 18808832 bytes (4%) 23494656 bytes (5%) 28196864 bytes (6%) 32899072 bytes (7%) 37601280 bytes (8%) 42303488 bytes (9%) 46989312 bytes (10%) 51691520 bytes (11%) 56393728 bytes (12%) 61095936 bytes (13%) 65781760 bytes (14%) 70483968 bytes (15%) 75186176 bytes (16%) 79888384 bytes (17%) 84590592 bytes (18%) 89276416 bytes (19%) 93978624 bytes (20%) 98680832 bytes (21%) 103383040 bytes (22%) 108068864 bytes (23%) 112771072 bytes (24%) 117473280 bytes (25%) 122175488 bytes (26%) 126877696 bytes (27%) 131563520 bytes (28%) 136265728 bytes (29%) 140967936 bytes (30%) 145670144 bytes (31%) 150372352 bytes (32%) 155058176 bytes (33%) 159760384 bytes (34%) 164455004 bytes (35%) 169165259 bytes (36%) 173859132 bytes (37%) 178561340 bytes (38%) 183262251 bytes (39%) 187949121 bytes (40%) 192651329 bytes (41%) 197345280 bytes (42%) 202047488 bytes (43%) 206749696 bytes (44%) 211451904 bytes (45%) 216137728 bytes (46%) 220839936 bytes (47%) 225542144 bytes (48%) 230244352 bytes (49%) 234946560 bytes (50%) 239632384 bytes (51%) 244334592 bytes (52%) 249036800 bytes (53%) 253739008 bytes (54%) 258441216 bytes (55%) 263127040 bytes (56%) 267829248 bytes (57%) 272531456 bytes (58%) 277233664 bytes (59%) 281919488 bytes (60%) 286621696 bytes (61%) 291323904 bytes (62%) 296026112 bytes (63%) 300728320 bytes (64%) 305414144 bytes (65%) 310116352 bytes (66%) 314818560 bytes (67%) 319520768 bytes (68%) 324206592 bytes (69%) 328908800 bytes (70%) 333611008 bytes (71%) 338313216 bytes (72%) 343015424 bytes (73%) 347701248 bytes (74%) 352403456 bytes (75%) 357105664 bytes (76%) 361807872 bytes (77%) 366510080 bytes (78%) 371195904 bytes (79%) 375898112 bytes (80%) 380600320 bytes (81%) 385302528 bytes (82%) 389988352 bytes (83%) 394690560 bytes (84%) 399392768 bytes (85%) 404094976 bytes (86%) 408797184 bytes (87%) 413483008 bytes (88%) 418185216 bytes (89%) 422887424 bytes (90%) 427589632 bytes (91%) 432275456 bytes (92%) 436977664 bytes (93%) 441679872 bytes (94%) 446382080 bytes (95%) 451084288 bytes (96%) 455770112 bytes (97%) 460472320 bytes (98%) 465174528 bytes (99%) 469864010 bytes (100%)
[ Info: loading pretrain gpt model: OpenAIftlm.npbson gpt_model
TransformerModel{Bert}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), segment = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512), postprocessor = Positionwise(LayerNorm(768), Dropout(0.1))),
  transformers = Bert(layers=12, head=12, head_size=64, pwffn_size=3072, size=768),
  classifier =
    (
      pooler => Dense(768 => 768, tanh)
      masklm => (
        transform => Chain(Dense(768 => 768, gelu), LayerNorm(768))
        output_bias => Vector{Float32}
      )
      nextsentence => Chain(Dense(768 => 2), logsoftmax)
    )
)TransformerModel{Gpt}(
  embed = CompositeEmbedding(tok = Embed(scale=1.0, 768), pe = PositionEmbedding(768, max_len=512)),
  transformers = Gpt(layers=12, head=12, head_size=64, pwffn_size=3072, size=768)
)

The pretrain"<type>-<model-name>:<item>" string with pretrain prefix will load the specific item from a known pretrain file (see the list below). The <type> is matched case insensitively, so not matter bert, Bert, BERT, or even bErT will find the BERT pretrain model. On the other hand, the <model-name>, and <item> should be exactly the one on the list. See example.

Currently support pretrain:

Type	model	model name	support items	detail description

Gpt	gpt	OpenAIftlm	gpt_model, bpe, vocab, tokenizer	gpt model with 12 layers, 768 hidden units, and 12 attention heads. subword with 40000-vocabularies-sized uncased bpe tokenizer.

Bert	scibert	scibert_scivocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer. vocabularies from scientific corpus.

Bert	scibert	scibert_basevocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	scibert	scibert_basevocab_uncased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	scibert	scibert_scivocab_cased	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer. vocabularies from scientific corpus.

Bert	bert	cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	wwm_cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	uncased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece uncased tokenizer.

Bert	bert	multi_cased_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multi_cased tokenizer.

Bert	bert	wwm_uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer. trained with whole word masked.

Bert	bert	multilingual_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece multilingual tokenizer.

Bert	bert	chinese_L-12_H-768_A-12	bert_model, wordpiece, tokenizer	bert model with 12 layers, 768 hidden units, and 12 attention heads. subword with google wordpiece chinese tokenizer.

Bert	bert	cased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece cased tokenizer.

Bert	bert	uncased_L-24_H-1024_A-16	bert_model, wordpiece, tokenizer	bert model with 24 layers, 1024 hidden units, and 16 attention heads. subword with google wordpiece uncased tokenizer.

If you don't find a public pretrain you want on the list, please fire an issue.

See example folder for the complete example.

API reference

Transformers.Pretrain.load_pretrain — Method

load_pretrain(name; kw...)

same as @pretrain_str, but can pass keyword argument if needed.

Transformers.Pretrain.pretrains — Function

pretrains(query::String = ""; detailed::Bool = false)

Show all available models. you can also query a specific model or model name. show more detail with detailed = true.

Transformers.Pretrain.@pretrain_str — Macro

pretrain"model-description:item"

convenient macro for loading data from pretrain. Use DataDeps to automatically download if a model is not found. the string should be in pretrain"<type>-<model-name>:<item>" format.

see also Pretrain.pretrains().